代理节点起步

+++ Assets/Script/Bridge/RuntimeData/ComponentData/Graph/Connection/PortConnectionData.cs
@@ -5,8 +5,8 @@ namespace YBFramework.Bridge
     [Serializable]
     public class PortConnectionData
     {
-        public ushort NodeID;
+        public int NodeID;
 
-        public ushort PortID;
+        public int PortID;
     }
 }
